[TM-1237] add hooks useHideOnScroll and useClickOutside
